// src/types/reserve.ts
export interface ReservationRequest {
  p02: P02[];
  p01: string;
  p03: '0';
}

interface P02 {
  r01: number;
  r02: number;
  r03: number;
  r04: number;
  r05: number;
  r06: number;
  r07: string;
}

// Generated by https://quicktype.io

export interface ReserveResponse {
  headers: { [key: string]: string };
  data: Data;
  status: number;
}

export interface Data {
  r01: string;
  r02: R02[];
}

export interface R02 {
  r01: string;
  r02: null;
  r03: number;
  r04: string;
  r05: number;
  r06: number;
  r07: number;
  r08: number;
  r09: string;
  r10: string;
  r11: boolean;
  r12: string;
}